refactor: NewRSACommand command has been changed to RSACommand
